requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
requests>=2.31.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
PyPDF2>=3.0.1
langchain-text-splitters>=0.0.1
beautifulsoup4>=4.12.0
lxml>=4.9.0
urllib3>=2.0.0